requirements.txt
fastapi
uvicorn
docker
websockets
jinja2
aiohttp
psutil
python-dateutil
PyJWT
python-multipart